Add support for CSS modules to extensions' compiler
