+++ bokehjs/make/package.json
@@ -12,12 +12,10 @@
     "./src/compiler"
   ],
   "devDependencies": {
-    "@types/css": "^0.0.33",
     "@types/node": "^18.7.16",
     "@types/which": "^2.0.1",
     "@types/yargs": "^16.0.4",
     "chalk": "^4.1.2",
-    "css": "^3.0.0",
     "del": "^6.1.1",
     "semver": "^7.3.7",
     "ts-node": "^10.9.1",
