+++ bokehjs/src/compiler/package.json
@@ -11,6 +11,7 @@
   "devDependencies": {
     "@types/combine-source-map": "^0.8.2",
     "@types/convert-source-map": "^1.5.2",
+    "@types/css": "^0.0.33",
     "@types/eslint": "^8.4.6",
     "@types/less": "^3.0.3",
     "@types/node": "^18.7.16",
@@ -21,6 +22,7 @@
     "chalk": "^4.1.2",
     "combine-source-map": "^0.8.0",
     "convert-source-map": "^1.8.0",
+    "css": "^3.0.0",
     "eslint": "^8.23.0",
     "eslint-plugin-deprecation": "^1.3.2",
     "less": "^4.1.3",
